slightly refactor graph editor into smaller chunks
