+++ app/gui2/src/providers/graphSelection.ts
@@ -1,12 +1,12 @@
 import type { NavigatorComposable } from '@/composables/navigator'
-import { useSelection } from '@/composables/selection'
+import { useSelection, type SelectionComposable } from '@/composables/selection'
 import { createContextStore } from '@/providers'
 import { type NodeId } from '@/stores/graph'
 import type { Rect } from '@/util/data/rect'
 
 const SELECTION_BRUSH_MARGIN_PX = 6
 
-export type GraphSelection = ReturnType<typeof injectFn>
+export type GraphSelection = SelectionComposable<NodeId>
 export { injectFn as injectGraphSelection, provideFn as provideGraphSelection }
 const { provideFn, injectFn } = createContextStore(
   'graph selection',
