+++ app/gui2/src/util/data/iterable.ts
@@ -1,1 +1,7 @@
 export * from 'shared/util/data/iterable'
+
+export function* filterDefined<T>(iterable: Iterable<T | undefined>): IterableIterator<T> {
+  for (const value of iterable) {
+    if (value !== undefined) yield value
+  }
+}
